cpp_python_integration
CMake files to create python library for C++ functions. New field calculation GUI to use these
