+++ DiWaCAT.bat
@@ -21,7 +21,39 @@ if %PIP_ERRORLEVEL% neq 0 (
     echo Failed to install some dependencies. Continuing anyway...
 )
 
+echo Attempting to Build Cpp Library
+cd Cpp_Source
+mkdir build
+cd build
+setlocal enabledelayedexpansion
+pause
+:: Find where pybind11 is installed and what architecture to build the python library to
+for /f %%i in ('python -c "import pybind11; print(pybind11.get_cmake_dir())"') do set PYBIND11_PATH=%%i
+for /f %%i in ('python -c "import struct; print(8 * struct.calcsize('P'))"') do set PYTHON_ARCH=%%i
+if %PYTHON_ARCH%==64 (
+    set CMAKE_ARCH=x64
+) else (
+    set CMAKE_ARCH=Win32
+)
+pause
+cmake -DCMAKE_BUILD_TYPE=Release -Dpybind11_DIR=!PYBIND11_PATH! -A !CMAKE_ARCH! ..
+if errorlevel 1 (
+    echo CMake configuration failed!
+	pause
+    exit /b 1
+)
+echo Building with CMake...
+cmake --build . --config Release
+if errorlevel 1 (
+    echo CMake build failed!
+	pause
+    exit /b 1
+)
+
+
+cd ../..
 echo Attempting to Run DiWaCAT
+python --version
 python DiWaCAT_GUI.py
 
 pause
