--- FieldSolver_Executable/Cpp_Source/DiWaCAT_FieldSolverUI.vcxproj.user
@@ -1,8 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="14.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'">
-    <LocalDebuggerEnvironment>PATH=%PATH%;$(TargetDir)\DLLS</LocalDebuggerEnvironment>
-    <DebuggerFlavor>WindowsLocalDebugger</DebuggerFlavor>
-    <LocalDebuggerMergeEnvironment>true</LocalDebuggerMergeEnvironment>
-  </PropertyGroup>
-</Project>
